feat: adds wiremock plugin.
