+++ core/config-resolver-wiremock/src/main/java/io/gatehill/imposter/plugin/wiremock/model/MappingResponse.kt
@@ -40,7 +40,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with Imposter.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.gatehill.imposter.config.resolver.model
+package io.gatehill.imposter.plugin.wiremock.model
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
