+++ core/config-resolver-wiremock/src/main/resources/META-INF/plugin.properties
@@ -1,0 +1,3 @@
+name=wiremock
+class=io.gatehill.imposter.plugin.wiremock.WiremockPluginImpl
+load=lazy
